Add `DeclarativeMetadata` concern
Similar to the `DeclarativeTouch` concern, this will allow us to trigger a
metadata refresh for a given object(s) off the back of model changes.
